let endpoints indicate they are unavailable
